requirements.txt
fastapi[standard]
pydantic
logfire
python-dotenv